requirements.txt
# Core dependencies
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
openai>=1.12.0
tiktoken>=0.5.0
numpy>=1.24.0
scipy>=1.11.0

# Vector stores and embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Optimization
ortools>=9.9.0
scikit-learn>=1.3.0

# Caching
redis>=5.0.0
cachetools>=5.3.0

# Bandit algorithms
mabwiser>=1.12.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
tenacity>=8.2.0
requests>=2.31.0

# Logging and monitoring
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

